+++ pkg/clusterfile/clusterfile.go
@@ -18,7 +18,8 @@ import (
 	"errors"
 	"sync"
 
-	"github.com/sealerio/sealer/pkg/runtime/kubernetes"
+	"github.com/sealerio/sealer/pkg/runtime/kubernetes/kubeadm"
+
 	v1 "github.com/sealerio/sealer/types/api/v1"
 	v2 "github.com/sealerio/sealer/types/api/v2"
 )
@@ -29,7 +30,7 @@ type ClusterFile struct {
 	path       string
 	Cluster    v2.Cluster
 	Configs    []v1.Config
-	KubeConfig *kubernetes.KubeadmConfig
+	KubeConfig *kubeadm.KubeadmConfig
 	Plugins    []v1.Plugin
 }
 
@@ -43,7 +44,7 @@ type Interface interface {
 	GetCluster() v2.Cluster
 	GetConfigs() []v1.Config
 	GetPlugins() []v1.Plugin
-	GetKubeadmConfig() *kubernetes.KubeadmConfig
+	GetKubeadmConfig() *kubeadm.KubeadmConfig
 }
 
 func (c *ClusterFile) GetCluster() v2.Cluster {
@@ -58,7 +59,7 @@ func (c *ClusterFile) GetPlugins() []v1.Plugin {
 	return c.Plugins
 }
 
-func (c *ClusterFile) GetKubeadmConfig() *kubernetes.KubeadmConfig {
+func (c *ClusterFile) GetKubeadmConfig() *kubeadm.KubeadmConfig {
 	return c.KubeConfig
 }
