+++ pkg/clusterfile/pre_process.go
@@ -20,10 +20,11 @@ import (
 	"io"
 	"os"
 
+	"github.com/sealerio/sealer/pkg/runtime/kubernetes/kubeadm"
+
 	"github.com/sealerio/sealer/common"
 	"github.com/sealerio/sealer/pkg/config"
 	"github.com/sealerio/sealer/pkg/env"
-	"github.com/sealerio/sealer/pkg/runtime/kubernetes"
 	v1 "github.com/sealerio/sealer/types/api/v1"
 	v2 "github.com/sealerio/sealer/types/api/v2"
 	yaml2 "github.com/sealerio/sealer/utils"
@@ -162,7 +163,7 @@ func (c *ClusterFile) DecodePlugins(data []byte) error {
 }
 
 func (c *ClusterFile) DecodeKubeadmConfig(data []byte) error {
-	kubeadmConfig, err := kubernetes.LoadKubeadmConfigs(string(data), yaml2.DecodeCRDFromString)
+	kubeadmConfig, err := kubeadm.LoadKubeadmConfigs(string(data), yaml2.DecodeCRDFromString)
 	if err != nil {
 		return err
 	}
