+++ pkg/runtime/kubernetes/reset.go
@@ -25,7 +25,7 @@ import (
 	"golang.org/x/sync/errgroup"
 )
 
-func (k *KubeadmRuntime) reset() error {
+func (k *Runtime) reset() error {
 	k.resetNodes(k.GetNodeIPList())
 	k.resetMasters(k.GetMasterIPList())
 	//if the executing machine is not in the cluster
@@ -41,7 +41,7 @@ func (k *KubeadmRuntime) reset() error {
 	return k.DeleteRegistry()
 }
 
-func (k *KubeadmRuntime) resetNodes(nodes []net.IP) {
+func (k *Runtime) resetNodes(nodes []net.IP) {
 	eg, _ := errgroup.WithContext(context.Background())
 	for _, node := range nodes {
 		node := node
@@ -57,7 +57,7 @@ func (k *KubeadmRuntime) resetNodes(nodes []net.IP) {
 	}
 }
 
-func (k *KubeadmRuntime) resetMasters(nodes []net.IP) {
+func (k *Runtime) resetMasters(nodes []net.IP) {
 	for _, node := range nodes {
 		if err := k.resetNode(node); err != nil {
 			logrus.Errorf("failed to delete master(%s): %v", node, err)
@@ -65,7 +65,7 @@ func (k *KubeadmRuntime) resetMasters(nodes []net.IP) {
 	}
 }
 
-func (k *KubeadmRuntime) resetNode(node net.IP) error {
+func (k *Runtime) resetNode(node net.IP) error {
 	ssh, err := k.getHostSSHClient(node)
 	if err != nil {
 		return fmt.Errorf("failed to reset node: %v", err)
