--- pkg/runtime/kubernetes/static_files.go
@@ -1,33 +1,0 @@
-// Copyright © 2021 Alibaba Group Holding Ltd.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package kubernetes
-
-const (
-	AuditPolicyYml = "audit-policy.yml"
-)
-
-// StaticFile :static file should not be template, will never be changed while initialization.
-type StaticFile struct {
-	DestinationDir string
-	Name           string
-}
-
-//MasterStaticFiles Put static files here, can be moved to all master nodes before kubeadm execution
-var MasterStaticFiles = []*StaticFile{
-	{
-		DestinationDir: "/etc/kubernetes",
-		Name:           AuditPolicyYml,
-	},
-}
